subscription_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
stripe==7.8.0
boto3==1.34.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
emails==0.6.0
jinja2==3.1.2
httpx==0.25.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0 